fix(popover): rename files for better gif diffs
